+++ modules/docker-build/main.tf
@@ -29,9 +29,7 @@ resource "docker_registry_image" "this" {
 
   keep_remotely = var.keep_remotely
 
-  triggers = {
-    image_id = docker_image.this.image_id
-  }
+  triggers = length(var.triggers) == 0 ? { image_id = docker_image.this.image_id } : var.triggers
 }
 
 resource "aws_ecr_repository" "this" {